update docs, tests, CI setup for new protoc method
Update docs, tests and CI setup to take advantage of the new protoc method that uses protoc_jll.
There is no need to install protoc separately and test against multiple versions of it.
CI on Windows is also enabled now.

+++ test/testcodec.jl
@@ -366,12 +366,12 @@ function test_types()
         end
 
         @testset "varint overflow" begin
-            ProtoBuf._write_uleb(pb, -1)
+            ProtoBuf._write_uleb(pb, Int64(-1))
             @test ProtoBuf._read_uleb(pb, Int8) == 0
-            ProtoBuf._write_uleb(pb, 1)
+            ProtoBuf._write_uleb(pb, Int64(1))
             @test ProtoBuf._read_uleb(pb, Int8) == 1
             write(pb, 0xff)
-            ProtoBuf._write_uleb(pb, -1)
+            ProtoBuf._write_uleb(pb, Int64(-1))
             @test ProtoBuf._read_uleb(pb, Int32) == 0
         end
